+++ pennylane/vqe/vqe.py
@@ -21,13 +21,7 @@ import pennylane as qml
 from pennylane.operation import Observable, Tensor
 
 
-OBS_MAP = {
-    "PauliX": "X",
-    "PauliY": "Y",
-    "PauliZ": "Z",
-    "Hadamard": "H",
-    "Identity": "I"
-}
+OBS_MAP = {"PauliX": "X", "PauliY": "Y", "PauliZ": "Z", "Hadamard": "H", "Identity": "I"}
 
 
 class Hamiltonian:
@@ -198,12 +192,15 @@ class VQECost:
     The cost function can be minimized using any gradient descent-based
     :doc:`optimizer </introduction/optimizers>`.
     """
+
     def __init__(self, ansatz, hamiltonian, device, interface="autograd", diff_method="best"):
         coeffs, observables = hamiltonian.terms
         self.hamiltonian = hamiltonian
         """Hamiltonian: the hamiltonian defining the VQE problem."""
 
-        self.qnodes = qml.map(ansatz, observables, device, interface=interface, diff_method=diff_method)
+        self.qnodes = qml.map(
+            ansatz, observables, device, interface=interface, diff_method=diff_method
+        )
         """QNodeCollection: The QNodes to be evaluated. Each QNode corresponds to the
         the expectation value of each observable term after applying the circuit ansatz.
         """