+++ tests/test_templates_utils.py
@@ -17,7 +17,7 @@ Tests for the templates utility functions.
 # pylint: disable=protected-access,cell-var-from-loop
 import pytest
 import numpy as np
-from pennylane.variable import Variable
+from pennylane.variable import VariableRef
 from pennylane.templates.utils import (_check_wires,
                                        _check_shape,
                                        _check_shapes,
@@ -27,7 +27,6 @@ from pennylane.templates.utils import (_check_wires,
                                        _check_is_in_options,
                                        _check_type)
 
-
 #########################################
 # Inputs
 
@@ -36,11 +35,13 @@ WIRES_PASS = [(0, [0]),
               ([1, 2], [1, 2])]
 WIRES_FAIL = [[-1],
               ['a'],
-              lambda x: x]
+              lambda x: x,
+              None
+              ]
 
 SHAPE_PASS = [(0.231, (), None),
               ([[1., 2.], [3., 4.]], (2, 2), None),
-              ([-2.3], (1, ), None),
+              ([-2.3], (1,), None),
               ([-2.3, 3.4], (4,), 'max'),
               ([-2.3, 3.4], (1,), 'min'),
               ([-2.3], (1,), 'max'),
@@ -50,14 +51,14 @@ SHAPE_PASS = [(0.231, (), None),
               ]
 
 SHAPE_LST_PASS = [([0.231, 0.1], [(), ()], None),
-                  ([[1., 2.], [4.]], [(2, ), (1, )], None),
-                  ([[-2.3], -1.], [(1, ), ()], None),
+                  ([[1., 2.], [4.]], [(2,), (1,)], None),
+                  ([[-2.3], -1.], [(1,), ()], None),
                   ([[-2.3, 0.1], -1.], [(1,), ()], 'min'),
                   ([[-2.3, 0.1], -1.], [(3,), ()], 'max')
                   ]
 
 SHAPE_FAIL = [(0.231, (1,), None),
-              ([[1., 2.], [3., 4.]], (2, ), None),
+              ([[1., 2.], [3., 4.]], (2,), None),
               ([-2.3], (4, 5), None),
               ([-2.3, 3.4], (4,), 'min'),
               ([-2.3, 3.4], (1,), 'max'),
@@ -66,37 +67,39 @@ SHAPE_FAIL = [(0.231, (1,), None),
               ]
 
 GET_SHAPE_PASS = [(0.231, ()),
+                  (complex(1, 0), ()),
+                  (1, ()),
                   ([[1., 2.], [3., 4.]], (2, 2)),
-                  ([-2.3], (1, )),
+                  ([-2.3], (1,)),
                   ([-2.3, 3.4], (2,)),
                   ([-2.3], (1,)),
                   ([[-2.3, 3.4, 1.], [1., 0.2, 1.]], (2, 3)),
                   ]
 
-#TODO: Think of a data structure that CANNOT be converted to numpy array
-GET_SHAPE_FAIL = []
+GET_SHAPE_FAIL = [("a",),
+                  (None,)]
 
-SHAPE_LST_FAIL = [([0.231, 0.1], [(), (3, 4)], None),
-                  ([[1., 2.], [4.]], [(1, ), (1, )], None),
+SHAPE_LST_FAIL = [([[0.231, 0.1]], [[()], [(3, 4)]], None),
+                  ([[1., 2.], [4.]], [(1,), (1,)], None),
                   ([[-2.3], -1.], [(1, 2), (1,)], None),
                   ([[-2.3, 0.1], -1.], [(1,), ()], 'max'),
                   ([[-2.3, 0.1], -1.], [(3,), ()], 'min')
                   ]
 
 LAYERS_PASS = [([[1], [2], [3]], 1),
-               ([[[1], [2], [3]], [['a'], ['b'], ['c']]], 3),
-             ]
+               ([[[1], [2], [3]], [[1], [2], [3]]], 3),
+               ]
 
-LAYERS_FAIL = [([1, 2, 3], None),
-               ([[[1], [2], [3]], [['b'], ['c']]], 3),
-              ]
+LAYERS_FAIL = [([[[1], [2], [3]], 1], 5),
+               ([[[1], [2], [3]], [[1], [2]]], 4),
+               ]
 
 NO_VARIABLES_PASS = [[[], np.array([1., 4.])],
                      [1, 'a']]
 
-NO_VARIABLES_FAIL = [[[Variable(0.1)], Variable([0.1])],
-                     np.array([Variable(0.3), Variable(4.)]),
-                     Variable(-1.)]
+NO_VARIABLES_FAIL = [[[VariableRef(0.1)], VariableRef([0.1])],
+                     np.array([VariableRef(0.3), VariableRef(4.)]),
+                     VariableRef(-1.)]
 
 OPTIONS_PASS = [("a", ["a", "b"])]
 
@@ -105,13 +108,14 @@ OPTIONS_FAIL = [("c", ["a", "b"])]
 TYPE_PASS = [(["a"], list, type(None)),
              (1, int, type(None)),
              ("a", int, str),
-             (Variable(1.), list, Variable)
+             (VariableRef(1.), list, VariableRef)
              ]
 
 TYPE_FAIL = [("a", list, type(None)),
-             (Variable(1.), int, list),
-             (1., Variable, type(None))
+             (VariableRef(1.), int, list),
+             (1., VariableRef, type(None))
              ]
+
 
 ##############################
 
@@ -138,7 +142,7 @@ class TestInputChecks:
 
     @pytest.mark.parametrize("wires", WIRES_FAIL)
     def test_check_wires_exception(self, wires):
-        """Tests that wires check fails if ``wires`` is not an integer or iterable."""
+        """Tests that wires check fails if ``wires`` is not a positive integer or iterable of positive integers."""
         with pytest.raises(ValueError, match="wires must be a positive integer"):
             _check_wires(wires=wires)
 
@@ -148,6 +152,12 @@ class TestInputChecks:
         shape = _get_shape(inpt)
         assert shape == target_shape
 
+    @pytest.mark.parametrize("inpt", GET_SHAPE_FAIL)
+    def test_get_shape_exception(self, inpt):
+        """Tests that ``_get_shape`` fails if unkown type of arguments."""
+        with pytest.raises(ValueError, match="could not extract shape of object"):
+            _get_shape(inpt)
+
     @pytest.mark.parametrize("inpt, target_shape, bound", SHAPE_PASS)
     def test_check_shape(self, inpt, target_shape, bound):
         """Tests that shape check succeeds for valid arguments."""
@@ -156,7 +166,7 @@ class TestInputChecks:
     @pytest.mark.parametrize("inpt, target_shape, bound", SHAPE_LST_PASS)
     def test_check_shape_list_of_inputs(self, inpt, target_shape, bound):
         """Tests that list version of shape check succeeds for valid arguments."""
-        _check_shapes(inpt, target_shape, bounds=[bound]*len(inpt), msg="XXX")
+        _check_shapes(inpt, target_shape, bounds=[bound] * len(inpt), msg="XXX")
 
     @pytest.mark.parametrize("inpt, target_shape, bound", SHAPE_FAIL)
     def test_check_shape_exception(self, inpt, target_shape, bound):
@@ -168,7 +178,7 @@ class TestInputChecks:
     def test_check_shape_list_of_inputs_exception(self, inpt, target_shape, bound):
         """Tests that list version of shape check succeeds for valid arguments."""
         with pytest.raises(ValueError, match="XXX"):
-            _check_shapes(inpt, target_shape, bounds=[bound]*len(inpt), msg="XXX")
+            _check_shapes(inpt, target_shape, bounds=[bound] * len(inpt), msg="XXX")
 
     @pytest.mark.parametrize("hp, opts", OPTIONS_PASS)
     def test_check_options(self, hp, opts):
@@ -200,7 +210,6 @@ class TestInputChecks:
 
     @pytest.mark.parametrize("inpt, repeat", LAYERS_FAIL)
     def test_check_num_layers_exception(self, inpt, repeat):
-        """Tests that layer check throws exception for invalid arguments."""
+        """Tests that layer check throws exception if number of layers not consistent."""
         with pytest.raises(ValueError, match="the first dimension of the weight parameters"):
             _check_number_of_layers(inpt)
-